backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.4
plotly==5.17.0
python-multipart==0.0.6
supabase==2.3.0
postgrest==0.13.0